Setup test DB using H2
Add the H2 dependency.
Use an in-memory H2 database.
Move the properties file to a better location and modify the path Config.
Add a persistence.xml file for testing.
Declare explicitly the entity classes in both persistence.xml files because of problems with H2.
Create a dummy test case for UsersFacade.
Add an initialization SQL script for H2.