Apply the new UI to the feedback creation, sign up and password recover forms.
Fix padding-right in the search bar.
Change the style of the search field in the navbar to make it look like the search bar.
Fix problems with bootstrap-responsive and nav-collapse
Set a fixed height for the header in the home page and the login form.
Implement a basic form of carousel in the header of the home page.
Separate some styles in various files.